Separate queries from charts
